Fix jersey basepath bug
